refactor(credential)!: keep store decorators in nebula-storage; drop reintroduced in-memory store double (ADR-0092 amend)
Step 3 (commit 04d179e5) moved the storage-coupled Encryption/Cache/Audit
decorators + KeyProvider into Core `nebula-credential` and, to test them there
(durable Sqlite stores are unreachable from Core), introduced an
`InMemoryCredentialStore` double. That reversed the deliberate durable-only
direction (#789 durable SQLite/Postgres, #790 deleting the in-memory store +
#788 deleting the testutil crate) and was a regression.

Correction:
- The `Encryption`/`Cache`/`Audit` decorators + `KeyProvider`/`AuditSink` stay in
  `nebula-storage`, restored with their original `SqliteCredentialStore`-backed
  tests (no behavioral change — original code/tests from `04d179e5^`).
- `nebula-credential::store_layer` is deleted; `InMemoryCredentialStore` is gone
  (zero refs). credential drops the `moka`/`nebula-env` deps + `tempfile` dev-dep.
- The `Cipher`/`Kdf` trait (in `nebula-crypto`) stays; wiring `EncryptionLayer`
  onto it is a storage-local follow-up, not a Core relocation.
- ADR-0092 amended: decorators are storage-coupled and durable-only-tested;
  they do not move into Core.

The other ADR-0092 moves (dispatchers/executor/scoped-accessor 4A.1, lease 4A.2,
fanout→resource) are unaffected. credential 292 tests green; storage 243 green
(decorator Sqlite integration tests restored).
