+++ crates/credential/Cargo.toml
@@ -62,12 +62,6 @@ semver = { workspace = true }
 # Synchronization
 arc-swap = { workspace = true }
 
-# Store-layer deps (ADR-0092 step 3: decorators relocated from nebula-storage)
-# moka: LRU+TTL cache used by CacheLayer; future-enabled for async invalidation.
-moka = { workspace = true, features = ["future"] }
-# nebula-env: EnvKeyProvider reads NEBULA_CRED_MASTER_KEY via nebula_env::var.
-nebula-env = { workspace = true }
-
 # Registry primitives (Tech Spec §3.1 — KEY-keyed CredentialRegistry)
 # `ahash` 0.8 — ~3× faster than default SipHash at credential-key sizes
 # (10–50 bytes); meets the §3.1 baseline bench (~5.5 ns mean lookup).
@@ -82,7 +76,6 @@ compact_str = { workspace = true }
 
 [dev-dependencies]
 base64 = { workspace = true }
-tempfile = { workspace = true }
 tokio = { workspace = true, features = [
   "io-util",
   "macros",
