+++ crates/storage/src/credential/layer/mod.rs
@@ -1,26 +1,19 @@
-//! ADR-0092 drain shim — deleted in step 8.
+//! Composable storage layers for [`CredentialStore`](nebula_credential::CredentialStore).
 //!
-//! The three decorator types and their supporting types were relocated to
-//! `nebula_credential::store_layer` (Core tier) in ADR-0092 step 3.
-//! These re-exports keep the previous `nebula_storage::credential` public
-//! surface resolving without changes at every call site.
+//! The multi-tenant **scope** layer was re-homed to the tenancy security
+//! boundary (`nebula_tenancy::CredentialScopeLayer` /
+//! `CredentialScopeResolver`, spec §8) — scope *policy* does not belong in
+//! the storage adapter. These layers (`EncryptionLayer`, `CacheLayer`,
+//! `AuditLayer`) stay here and re-compose **on top** of the tenancy scope
+//! layer at the composition root; the fail-closed audit +
+//! zeroize-on-drop invariants are unaffected by the move (layer order
+//! `ScopeLayer → AuditLayer → EncryptionLayer → CacheLayer → Backend` is
+//! preserved by the composition root, not by this module).
 
-/// Drain shim — see [`nebula_credential::store_layer::audit`].
-pub mod audit {
-    pub use nebula_credential::store_layer::audit::*;
-}
+pub mod audit;
+pub mod cache;
+pub mod encryption;
 
-/// Drain shim — see [`nebula_credential::store_layer::cache`].
-pub mod cache {
-    pub use nebula_credential::store_layer::cache::*;
-}
-
-/// Drain shim — see [`nebula_credential::store_layer::encryption`].
-pub mod encryption {
-    pub use nebula_credential::store_layer::encryption::*;
-}
-
-pub use nebula_credential::store_layer::{
-    AuditEvent, AuditLayer, AuditOperation, AuditResult, AuditSink, CacheConfig, CacheLayer,
-    CacheStats, EncryptionLayer,
-};
+pub use audit::{AuditEvent, AuditLayer, AuditOperation, AuditResult, AuditSink};
+pub use cache::{CacheConfig, CacheLayer, CacheStats};
+pub use encryption::EncryptionLayer;
